Refactor config inititialization
Move initializing home dirs to a function in commands and call it in a
main function before parse_args(). Seperating the creation of the dirs
allows for easier testing.

SETTINGS will load default SettingsSpec() if CONFIG_FILE is not a file.
A new settings can be loaded for testing fairly easily while maintaining
functional default behavior.